requirements.txt
fastapi==0.93.0
replicate==0.5.1
python-multipart==0.0.6
uvicorn==0.21.0
Pillow==9.4.0
